+++ pages/explore/actions/AddAction.js
@@ -3,11 +3,11 @@ import styled from 'styled-components';
 import getConfig from 'next/config';
 import _ from 'lodash';
 
-import { useAuth } from '../../lib/Auth';
-import { useMatrix } from '../../lib/Matrix';
-import CreateContext from './createContext';
-import TemplateSelect from './templateSelect';
-import AddExistingContext from './AddExistingContext';
+import { useAuth } from '../../../lib/Auth';
+import { useMatrix } from '../../../lib/Matrix';
+import CreateContext from '../CreateContext';
+import TemplateSelect from '../TemplateSelect';
+import AddExistingContext from '../AddExistingContext';
 
 const ModSection = styled.div`
   & {
