+++ pages/explore/actions/InfoAction.js
@@ -3,9 +3,9 @@ import styled from 'styled-components';
 import getConfig from 'next/config';
 import _ from 'lodash';
 
-import { useAuth } from '../../lib/Auth';
-import { useMatrix } from '../../lib/Matrix';
-import UserHandle from './userHandle';
+import { useAuth } from '../../../lib/Auth';
+import { useMatrix } from '../../../lib/Matrix';
+import UserHandle from '../UserHandle';
 
 const InfoSection = styled.div`
   & {
