+++ pages/explore/actions/index.js
@@ -3,12 +3,12 @@ import styled from 'styled-components';
 import getConfig from 'next/config';
 import _ from 'lodash';
 
-import { useAuth } from '../../lib/Auth';
-import { useMatrix } from '../../lib/Matrix';
-import SettingsAction from './settingsAction';
-import InfoAction from './infoAction';
-import AddAction from './addAction';
-import RemoveAction from './removeAction';
+import { useAuth } from '../../../lib/Auth';
+import { useMatrix } from '../../../lib/Matrix';
+import SettingsAction from './SettingsAction';
+import InfoAction from './InfoAction';
+import AddAction from './AddAction';
+import RemoveAction from './RemoveAction';
 
 const ActionsSection = styled.details`
   & {
